pyproject.toml
[tool.poetry]
name = "project"
version = "0.2.0"
description = ""
authors = ["CuasarN1 <55405975+CuasarN1@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.2"
pydantic-settings = "^2.5.2"
uvicorn = "^0.31.1"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
alembic = "^1.13.3"
tenacity = "^9.0.0"
psycopg2 = "^2.9.10"
pyjwt = "^2.10.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.17"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"